Distributed System Load Monitor/requirements.txt
flask==2.0.1
psutil==5.8.0
requests==2.26.0
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
python-socketio[client]==5.4.0 